+++ server.js
@@ -5,7 +5,13 @@ const path = require('path')
 const domapic = require('domapic-service')
 const gpioOut = require('gpio-out-domapic')
 
-const { GPIO, INITIAL_STATUS, INVERT, REMEMBER_LAST_STATUS } = require('./lib/statics')
+const {
+  GPIO,
+  INITIAL_STATUS,
+  INVERT,
+  REMEMBER_LAST_STATUS,
+  PRESS_TIME
+} = require('./lib/statics')
 const options = require('./lib/options')
 
 const pluginConfigs = require('./lib/plugins')
@@ -14,6 +20,7 @@ domapic.createModule({
   packagePath: path.resolve(__dirname),
   customConfig: options
 }).then(async dmpcModule => {
+  const config = await dmpcModule.config.get()
   const relay = new gpioOut.Gpio(dmpcModule, {}, {
     gpio: GPIO,
     initialStatus: INITIAL_STATUS,
@@ -35,7 +42,7 @@ domapic.createModule({
         handler: () => relay.status
       },
       action: {
-        description: 'Switch on/off the relay',
+        description: 'Switches on/off the relay',
         handler: async (newStatus) => {
           await relay.setStatus(newStatus)
           dmpcModule.events.emit('switch', newStatus)
@@ -44,12 +51,26 @@ domapic.createModule({
       }
     },
     toggle: {
-      description: 'Toggle the relay status',
+      description: 'Toggles the relay status',
       action: {
-        description: 'Toggle the relay status',
+        description: 'Toggles the relay status',
         handler: async () => {
           const newStatus = await relay.toggle()
           dmpcModule.events.emit('switch', newStatus)
+          return Promise.resolve()
+        }
+      }
+    },
+    shortPress: {
+      description: 'Inverts the relay status during a defined period of time',
+      action: {
+        description: 'Inverts the relay status briefly',
+        handler: async () => {
+          const currentStatus = relay.status
+          await relay.toggle()
+          setTimeout(() => {
+            relay.setStatus(currentStatus)
+          }, config[PRESS_TIME])
           return Promise.resolve()
         }
       }