+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 
 ## [1.2.0] - 2018-12-25
+### Added
+- Add "shortPress" ability, which inverts the relay status during a defined period of time.
 
 ## [1.1.0] - 2018-12-22
 ### Added